+++ lib/generators/settings/install_generator.rb
@@ -26,7 +26,8 @@ module Settings
     end
 
     def install_setting
-      template 'model.rb', File.join('app/models', class_path, "#{file_name}.rb"), force: true
+      template 'model.rb', File.join('app/models', class_path, "#{file_name}.rb")
+      template 'app.yml', File.join('config', "app.yml")
       migration_template 'migration.rb', 'db/migrate/create_settings.rb'
     end
   end