+++ spec/rails-settings-cached/setting_spec.rb
@@ -88,8 +88,16 @@ describe RailsSettings do
     it 'should work' do
       expect(Setting.all.count).to eq 8
     end
+  end
+
+  describe '#get_all' do
+    it "should include defaults" do
+      expect(RailsSettings::Default).to receive(:instance).and_return({ default1: 1, default2: '2' })
+      expect(Setting.get_all).to include(:default1, :default2)
+    end
 
     it "should all('namespace')" do
+      expect(Setting.get_all('config')).to eq({ "config.color" => :red, "config.limit" => 100 })
       expect(Setting.get_all('config').count).to eq 2
     end
   end
@@ -109,18 +117,7 @@ describe RailsSettings do
     end
   end
 
-  describe 'Default values' do
-    it 'can work with default value' do
-      Setting.defaults[:bar] = @bar
-      expect(Setting.bar).to eq @bar
-    end
-
-    it 'can use default value, when the setting it cached with nil value' do
-      Setting.has_cached_nil_key
-      Setting.defaults[:has_cached_nil_key] = '123'
-      expect(Setting.has_cached_nil_key).to eq '123'
-    end
-
+  describe 'Save Default values' do
     it '#save_default' do
       Setting.test_save_default_key
       Setting.save_default(:test_save_default_key, '321')